app.py
# app.py
from flask import Flask, request, jsonify
import google.generativeai as genai
import os
# --- Import your existing Memory class and other logic here ---
# (You would copy/paste your Memory class and other helper functions)
# (Remove all tkinter, sounddevice, pyttsx3, and subprocess calls)

app = Flask(__name__)

# --- Initialize your memory and Gemini model here ---
# memory = Memory() 
# genai.configure(api_key=os.getenv("GEMINI_API_KEY"))
# model = genai.GenerativeModel(...)
# chat_session = model.start_chat(...)
# ----------------------------------------------------

@app.route('/chat', methods=['POST'])
def chat():
    data = request.get_json()
    user_message = data.get('message')

    if not user_message:
        return jsonify({"error": "No message provided"}), 400

    # --- Your existing command handling logic goes here ---
    # This is where you'd check for "weather", "search for", etc.
    # For this example, we'll just send it to Gemini.
    try:
        # response = chat_session.send_message(user_message)
        # response_text = response.text
        
        # Placeholder response for demonstration
        response_text = f"You said: {user_message}. The web version is still under construction."
        
        # memory.add_to_history("user", user_message)
        # memory.add_to_history("assistant", response_text)

        return jsonify({"reply": response_text})

    except Exception as e:
        print(f"Error: {e}")
        return jsonify({"error": "Failed to get a response from the AI."}), 500

if __name__ == '__main__':
    app.run(debug=True) # For local testing 